requirements.txt
torch>=2.0.0
transformers>=4.35.0
peft>=0.6.0
datasets>=2.14.0
wandb>=0.15.0
accelerate>=0.24.0
bitsandbytes>=0.41.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.21.0
scikit-learn>=1.0.0